"8:40am. https://youtu.be/V10hzjgoklA?t=2296
He got the stuff on this slide wrong. On the left side the `*` should be `+` and vice versa.

8:50am. https://youtu.be/V10hzjgoklA?t=2829

There is no way this second line is true. What the hell?

8:55am. https://youtu.be/V10hzjgoklA?t=2961

"Life is cool. Do something cool and stick to it for 30 years. And always remember that when you have a tough job to do that you should recall that this is a job...FOR LAMBDA CALCULUS!" (rips his short to reveal the LambdaMan costume underneath)

9am. That was fun, but everything in the talk is something that was already in the PLFA book. Based on this talk, it does not seem like I need to study category theory since all the stuff here is a part of functional programming anyway.

Let me go to the next thing on the list which are the two Coq exercises I am yet to do. I want to the Indprop chapter today. That will free up my mental space to move to the rest. If the rest of the book does not turn out balls to the wall hard as this chapter, I should be able to finish it by the time of the half yearly review.

Let me get this done. In a few months I will be done with all this preliminary work and will be able to do some real math that I am interested in. After I am done with the first SF volume, I am really interested in giving Cubical TT a shot. Maybe rather than Coq, I will end up using Agda in the end. Translating that math library from Coq to Agda would be excellent practice.

I'll have a lot of work to go through before I can get to the parts I am interested in. This is the same as when I worked on Spiral. I will get through this and when I am done nothing will stop me.

I really hope that if it has to be 30 years, that I will get thrown a bone every now and then. I really will end up on the street if nothing changes long before that if I can't get some kind of edge from all of this."