"5:20pm. https://stackoverflow.com/questions/41568683/why-is-it-impossible-to-perform-induction-on-a-term-that-is-used-in-conclusion
This is so complicated. I do not at all feel like studying what should be Coq internals. Agda can do it so why can't Coq?

5:30pm.

```
Theorem In_to_In' :
  forall {A : Type} (x : A) (l : list A),
  In x l -> In' x l.
```

Ah, I get what he meant. When I define it like this it typechecks.

```
Theorem In_to_In' :
  forall {A : Type} (x : A) (l : list A),
  In x l -> In' x l.
Proof.
intros.
induction l.
+ inversion H.
+ simpl in H.
```
```
1 subgoal
A : Type
x, x0 : A
l : list A
H : x0 = x \/ In x l
IHl : In x l -> In' x l
______________________________________(1/1)
In' x (x0 :: l)
```
```
  destruct H; subst.
```
```
Case analysis on sort Type is not allowed for inductive definition or.
```

Ahhhh...Yeah, I can't deal with this.

5:30pm. Forget this.

https://softwarefoundations.cis.upenn.edu/lf-current/Maps.html

Let me move to the Maps chapter.

```
(** One small digression before we begin...

    Unlike the chapters we have seen so far, this one does not
    [Require Import] the chapter before it (and, transitively, all the
    earlier chapters).  Instead, in this chapter and from now, on
    we're going to import the definitions and theorems we need
    directly from Coq's standard library stuff.  You should not notice
    much difference, though, because we've been careful to name our
    own definitions and theorems the same as their counterparts in the
    standard library, wherever they overlap. *
```

Oh, nice.

5:50pm. I've read the chapter. I definitely do not feel like going through the proofs today, but I should be able to do them all tomorrow.

https://softwarefoundations.cis.upenn.edu/lf-current/ProofObjects.html

Let me also scout this one. The Map chapter is pretty easy and I am interested in whether that is a change in the tides.

5:55pm. Ah, I had no idea it was possible to use tactics in definitions.

6:10pm. https://softwarefoundations.cis.upenn.edu/lf-current/ProofObjects.html

Had to take a little break. There is a 5-star exercise here to show that the inductive definition of equality is equal to the Leibnitz one, but that was covered in the Agda book so I should not have too much difficulty with it. I have the bits of its structure in mind so it should make for a decent review.

I think I should be able to do two chapters tomorrow. Or at least one per day.

There is not much in terms of exercises in either of them.

https://softwarefoundations.cis.upenn.edu/lf-current/IndPrinciples.html

1-star exercises here only.

6:15pm. https://softwarefoundations.cis.upenn.edu/lf-current/Imp.html

This one will be a bit harder. It has a bunch of 3 and 4 star exercises, some of the advanced.

https://softwarefoundations.cis.upenn.edu/lf-current/ImpParser.html

This one has no exercises.

...It seems all the really difficult stuff was in the IndProp chapter. Only `Imp` seems like it will be remotely challenging out of what remains. I should be able to finish it at a fast clip.

Great, that is great!

I guess I'll go for PL fundamentals after that even though I am not particularly interested in that material.

6:20pm. https://stackoverflow.com/questions/56415849/how-to-implement-remove-with-a-membership-proof-as-an-argument-in-coq/56415972?noredirect=1#comment99428788_56415972

I'll try proving `In x l <-> In' x l` according to this tomorrow so I'll accept the answer only after I've verified it. Right now I am too tired for this.

6:25pm. I am done for the day here.

Let me watch a bit on cubical type theory and then I'll get to resting."