"3:15pm. I am back from the bath. And in the fridge a cake was for me. Today is just so wonderful.
After the long struggle of the past month, I finally see the finish line to the first SF volume.

3:20pm. Before I can start that, let me see if Allais' advice is right on the mark.

3:30pm.

```
Goal: to' (from (to n)) ≡ ⟨ to n , Can- prf ⟩
————————————————————————————————————————————————————————————
prf' : to' (from (to n)) ≡ to' n
```

The two proofs are not equal apparently.

```
    to∘from : (a : Bin) → (b : Can a) → to' (from a) ≡ ⟨ a , b ⟩
    to∘from t (Can- {n} prf) =
      let prf' =  (Eq.subst (λ v → to' v ≡ to' n) (Eq.sym prf) {!Eq.sym prf!})
      in {!prf'!}
```

Why did I make the type signature so strange?

Let me do it normally.

3:35pm.

```
    to∘from : (x : ∃[ x ] Can x ) → to' (from' x) ≡ x
    to∘from ⟨ .(to n) , Can- {n}  x ⟩ = {!!}
```

When I do it like this the type error goes away. I see...

3:40pm.

```
Goal: to' (from (to n)) ≡ ⟨ to n , Can- x ⟩
————————————————————————————————————————————————————————————
x : from (to n) ≡ n
n : ℕ
```

The rewrite still fails on this...

Hmmmm...

3:45pm.

```
    to∘from ⟨ .(to n) , Can- {n} x ⟩ = Eq.subst {!λ v → to' v ≡ ⟨ to n , Can- x ⟩!} x {!!}
```

I really do not understand why I can't instantiate the hole with this.

3:45pm.

```
E:\TheoremProving\plfa\Quantifiers.lagda:597,96-100
from-to-identity != x of type (from (to n) ≡ n)
when checking that the expression refl has type
to' n ≡ ⟨ to n , Can- x ⟩
```

Hmmm...

Wasn't there one proof in which I destructured x into an actual function?

```
from-to-identity : ∀ {n : ℕ} -> from (to n) ≡ n
from-to-identity {zero} = refl
from-to-identity {suc n} rewrite from-helper-distr 1 (to n) | from-to-identity {n} = refl
```

I am betting that all of this is happening because of implicit arguments.

3:55pm.

```
E:\TheoremProving\plfa\Quantifiers.lagda:598,93-97
(plfa.Quantifiers.from-to-identity n) != from-to-identity of type
(from (to n) ≡ n)
when checking that the expression refl has type
to' n ≡ ⟨ to n , Can- from-to-identity ⟩
```

This is trouble.

4pm.

```
E:\TheoremProving\plfa\Quantifiers.lagda:598,84-88
(from-to-identity n) != prf of type (from (to n) ≡ n)
when checking that the expression refl has type
to' n ≡ ⟨ to n , Can- prf ⟩
```

After thinking about this for a bit, I see why Coq cannot see this even though it obvious to me. I guess a proof by induction should do the trick in that case.

```
    to∘from ⟨ .(to n) , Can- {n} prf ⟩ =
      Eq.subst (λ v → to' v ≡ ⟨ to n , Can- prf ⟩) (Eq.sym prf) {!refl!}
```

Let me back this up here for now.

```
    to∘from ⟨ .(x0 nil) , Can- {zero} prf ⟩ rewrite prf = refl
```

Here it is.

4:05pm.

```
    to∘from ⟨ .(x0 nil) , Can- {zero} prf ⟩ rewrite prf = refl
    to∘from ⟨ .(inc (to n)) , Can- {suc n} prf ⟩ = {!!}
```

This would be so much easier if the rewrite worked.

```
Goal: to' (from (inc (to n))) ≡ ⟨ inc (to n) , Can- prf ⟩
————————————————————————————————————————————————————————————
prf : from (inc (to n)) ≡ suc n
n   : ℕ
```

I should be able to get that `suc` out of `prf` in order to make the recursive call. In fact, I remember a proof to that effect.

But first, I need a partial rewrite.

4:20pm.

```
    to∘from ⟨ .(inc (to n)) , Can- {suc n} prf ⟩ with subst (λ v → {!!} ≡ suc n) (from-inc≡suc-from (to n)) refl
    ... | prf' = {!!}
```

I simply do not understand how I could relax this much. It is not like I've been slacking, but still the time just slipped through my fingers today.

And now I cannot do the subst here properly for some reason. I am going crazy.

I know I used subst successfully once before, but now it is like my understanding of it evaporated.

What exactly is going on here?"