"```         Can-iso prf with Can-suc (Can-pred prf)         ... | prf' = {!!} ``` ``` Goal: prf' ≡ prf ———————————————————————————————————————————————————————————— prf  : from-helper 1 (inc (to n₁)) ≡ suc n₁ x    : from-helper 1 (to n) ≡ n n    : ℕ prf' : from-helper 1 (inc (to n₁)) ≡ suc n₁ n₁   : ℕ  (not in scope) ```
As it turns out for that bag problem there was no need to separate it into steps. I opened an issue on the Agda repo once, and got completely mislead.

...No I can't case split on `prf` and `prf'` for some reason. Let me then try `refl-eq`.

```
        Can-iso prf with Can-suc (Can-pred prf)
        ... | prf' = refl-eq prf' prf
```

Yeah, this works.

It might be worth asking about this.

11:15am.

```
        loop-suc' : ∀ {n} {prf : from (to n) ≡ n } → to' n ≡ ⟨ to n , Can- prf ⟩ → to' (suc n) ≡ ⟨ to (suc n) , Can- (Can-suc prf) ⟩
        -- loop-suc' = cong λ { ⟨ _ , Can- {n} prf ⟩ → ⟨ to (suc n) , Can- (Can-suc prf) ⟩ }
```

Forget about this for the time being.

11:20am.

```
    to∘from ⟨ .(to n) , Can- {n} x ⟩ = loop n x
      where
        Can-pred : ∀ {n} → from (to (suc n)) ≡ suc n → from (to n) ≡ n
        Can-pred {n} _ = from-to-identity n
        Can-suc : ∀ {n} → from (to n) ≡ n → from (to (suc n)) ≡ suc n
        Can-suc {n} _ = from-to-identity (suc n)
        Can-iso : ∀ {n} (prf : from (to (suc n)) ≡ suc n) → Can-suc (Can-pred prf) ≡ prf
        Can-iso prf with Can-suc (Can-pred prf)
        ... | prf' = refl-eq prf' prf

        loop-suc' : ∀ {n} {prf : from (to n) ≡ n } → to' n ≡ ⟨ to n , Can- prf ⟩ → to' (suc n) ≡ ⟨ to (suc n) , Can- (Can-suc prf) ⟩
        -- loop-suc' = cong λ { ⟨ _ , Can- {n} prf ⟩ → ⟨ to (suc n) , Can- (Can-suc prf) ⟩ }

        loop-suc : ∀ {n} {prf : from (to n) ≡ n } → to' (from (to n)) ≡ ⟨ to n , Can- prf ⟩ → to' (suc (from (to n))) ≡ ⟨ to (suc n) , Can- (Can-suc prf) ⟩
        loop-suc {n} {prf} x = subst {!!} prf {!x!}

        loop : ∀ n (x : from (to n) ≡ n) → to' (from (to n)) ≡ ⟨ to n , Can- x ⟩
        loop zero prf rewrite prf = refl
        loop (suc n) prf =
          to' (from (to (suc n)))
            ≡⟨ cong to' (from-inc≡suc-from (to n)) ⟩
          to' (suc (from (to n)))
            ≡⟨ loop-suc (loop n (Can-pred prf)) ⟩
          ⟨ to (suc n) , Can- (Can-suc (Can-pred prf)) ⟩
            ≡⟨ cong (λ prf → ⟨ to (suc n) , Can- prf ⟩ ) (Can-iso prf) ⟩
          ⟨ to (suc n) , Can- prf ⟩
            ∎
```

I got a bit further, but this is just as impossible as before. The typechecker is preventing me what I need to do.

I am going to have to ask about this. No other way. Let make an update to that closed question.

11:35am.

```
Goal: prf-from ≡ there-prf
————————————————————————————————————————————————————————————
ys = ys₁  : List A₁  (not in scope)
xs = xs₁  : List A₁  (not in scope)
P = P₁    : A₁ → Set  (not in scope)
A = A₁    : Set  (not in scope)
prf-from  : Any P (x ∷ xs ++ ys)
x₁        : P x
there-prf : Any P (x ∷ xs ++ ys)
```

No wait, these two are not equalities. I am dumb.

Forget this. Good thing I did not make an extra comment and decided to check first.

11:45am. Let me take that break here. I'll compose a post summarizing my issues after I get back."