"9:40am. https://softwarefoundations.cis.upenn.edu/lf-current/IndPrinciples.html
This is the next chapter.

9:45am.

```
(** **** Exercise: 2 stars, optional (plus_one_r')  *)
(** Complete this proof without using the [induction] tactic. *)

Theorem plus_one_r' : forall n:nat,
  n + 1 = S n.
Proof.
apply nat_ind; intros.
- reflexivity.
- simpl. rewrite H. reflexivity.
Qed.
```

9:50am. Hmmmm...

```
Check nat_ind.
(*  ===> nat_ind :
           forall P : nat -> Prop,
              P 0  ->
              (forall n : nat, P n -> P (S n))  ->
              forall n : nat, P n  *)
```

I never realized it before, but is the book trying to highlight a connection between induction principles and Leibnitz substitution? Leibnitz substitution is induction for the equality type?

I am still groggy. I can't really work out the whether this is true or not right now, but I guess it might be the case. I'll add to the list of things to figure out.

```
Lemma equality__leibniz_equality : forall (X : Type) (x y: X),
  x = y -> forall P:X->Prop, P x -> P y.
```

Ah, this is literally `subst`. I must be asleep.

So no, it is not the induction principle for `x = y`, but the correspondence between the two.

10am.

```
(** **** Exercise: 1 star, optional (byntree_ind)  *)
(** Write out the induction principle that Coq will generate for the
    following datatype.  (Again, write down your answer on paper or
    type it into a comment, and then compare it with what Coq
    prints.) *)

Inductive byntree : Type :=
 | bempty : byntree
 | bleaf  : yesno -> byntree
 | nbranch : yesno -> byntree -> byntree -> byntree.
```

I'll skip these for the time being.

10:05am.

```
(** **** Exercise: 1 star, optional (ex_set)  *)
(** Here is an induction principle for an inductively defined
    set.

      ExSet_ind :
         forall P : ExSet -> Prop,
             (forall b : bool, P (con1 b)) ->
             (forall (n : nat) (e : ExSet), P e -> P (con2 n e)) ->
             forall e : ExSet, P e

    Give an [Inductive] definition of [ExSet]: *)

Inductive ExSet : Type :=
  | con1 : bool -> ExSet
  | con2 : nat -> ExSet -> ExSet.

Check ExSet_ind.
```

I decided to do this one. Ok.

10:15am. https://existentialtype.wordpress.com/2017/11/04/sequentiality-is-the-essence-of-parallelism/
```
So, what are the join points?  They are the elimination forms for two parallel modalities.  They generalize the sequential case to allow for statically and dynamically determined parallelism.   A value of parallel product type is a tuple of unevaluated computations, a kind of “lazy” tuple (but not that kind of laziness, here I just mean unevaluated components).  The elimination form evaluates all of the component computations in parallel, creates a value tuple from their values, and passes it to the body of the form.  Similarly, a value of parallel sequence type is a generator consisting of two values, a natural number n indicating its size, and a function determining the ith component computation for each 1≤i<n.  The elimination form activates all n component computations, binds their values to a value sequence, and passes it to the body of the form.
```

Does this have much to do with Spiral's join points? I guess I'll figure it out at some point.

Let me finish the chapter.

I am already 2/3rds into it.

10:30am. Focus me, stop surfing /a/.

...Ah, it is done. Let me move to the next chapter. That is 3 chapters down, just in one morning.

https://softwarefoundations.cis.upenn.edu/lf-current/Rel.html

The next one is this.

10:35am. It a short and optional chapter. Let me see if I can do it before breakfast.

10:45am.

```
(** **** Exercise: 2 stars, optional (total_relation_not_partial)  *)
(** Show that the [total_relation] defined in earlier is not a partial
    function. *)

(* FILL IN HERE *)
(** [] *)
```

While this chapter is short, it is not so short that I will be able to do it in 5m. It will take me a few hours.

Let me do this exercise and then I will take a break.

```
(** **** Exercise: 2 stars, optional (total_relation)  *)
(** Define an inductive binary relation [total_relation] that holds
    between every pair of natural numbers. *)

Inductive total_relation : nat -> nat -> Prop :=
  | tr : forall n m, total_relation n m.
```

Ah, it is the same as the new definition.

11:10am. I'll do it later. Let me have breakfast now."