"11:55am. https://mangadex.org/title/275/knight-run
Oh, more of this is out. It is really great that translations for this are coming at a steady clip. I had not expected it would be like this.

12:10pm. Let me resume. To to put another chapter away. The `Imp` chapter after that has a bunch of 3-4 star exercises. That one will be hard.

I am still groggy though. I had not gone to bed that late, so shouldn't I be more refreshed by now?

Well, no matter. Let me do my best.

I'll finish the current chapter and then think of what comes next after that.

12:20pm. Well then. It is time to start. I need to stop daydreaming and focus on the task at hand.

I know that I am fatigued, but I need to push on."