Isekai Quartet is really a lot better than it has any right to be. It actually makes me laugh from time to time.
...I am not really motivated at the moment. Lately during my free time, rather than playing games or watching anime I've just been consuming meaningless social media. I really made a mistake past high school when I started to consume news, rather than be the old ambivalent me. I consumed too much political tripe as a part of that.

The thing is, from pretty much every perspective except the cause-the-Singularity one, the activities I do now during my free times are worse than when I was a kid. The new habits I've ingrained have mostly caused me to degenerate as a person.

But ultimately I cannot afford the old habits anymore. The energy I would have put into games, is now going into studying and programming. I made a deal that if I could not hack it as a trader that I would give into my insanity - and bring this world to its knees.

The new habits while low, are cheap enough to sustain me.

If this quest stretches out for over a decade, I am not sure where I will find the will to do what is needed. Probably the same place as in the last 4.5 years.

Until the fundamentals change, neither will I.

Let me resume.

The next chapter is `ImpParser`. If I remember, that one probably does not have any exercises. In fact I remember that out of the 4 chapters that are left, only one has a bunch of 1-2 star exercises.

Unless I am wrong, I should be able to blow through them in the next few hours.