"8:25am. I am up.
Let me chill for a while and then I will start.

I feel like I understand the problem, but I lack the basic insight on how to do the connections, so I am going to try proving a few simple lemmas just to get a lay of the land. I admit that I did not want to do these problems, but now I am into it. I guess I'll spend a while trying to get through them.

9am. Let me finally start this thing. I am done with resting.

I am not sure how today will go. Every time I get stuck there is a possibility that I might not be able to overcome it. I might solve this with a few hours or effort, or it might take me days. There is always an element of gambling in this.

But I have no choice, but to believe in myself and just go forward.

That is what I've always done and what I will do once more."