Let me finally start the day. I am done chilling.
Today the plan is to do as I've discussed yesterday and try to do the problem in Agda. Coq is proving too hard for me when it comes to this.

These advanced problems are no joke.

I will try to aim to get at least the one I started 2 days ago done.

But no pressure. I will just keep plugging at this until I get it done.