"1:10pm.
"The book develops two main conceptual threads:
(1) formal techniques for reasoning about the properties of specific programs (e.g., the fact that a sorting function or a compiler obeys some formal specification); and
(2) the use of type systems for establishing well-behavedness guarantees for all programs in a given programming language (e.g., the fact that well-typed Java programs cannot be subverted at runtime)."

I really am interested in the first one.

Damn, I am going to have to go through this as well.

Well, maybe it won't be as hard as the first part.

1:15pm. "The Coq proofs we're doing now are sufficiently complicated that it is more or less impossible to complete them by random experimentation or "following your nose." You need to start with an idea about why the property is true and how the proof is going to go. The best way to do this is to write out at least a sketch of an informal proof on paper — one that intuitively convinces you of the truth of the theorem — before starting to work on the formal one. Alternately, grab a friend and try to convince them that the theorem is true; then try to formalize your explanation."

1:25pm. The second volume is enormous.

...I am going to be working on this for the next few months at this rate...

Agh, but I do need this. I need these exercises if I am going to do the efficient Bayesian inference algorithm for NNs.

1:30pm. I am going to have to do it. I am going to do vol 2 in Agda and then in Coq, before moving on to volume 3. I guess that will be my goal for the next 3 months.

If I want to do real math with dependent types, then this material is valuable enough as exercise that I have to go through this.

It would be different if it was the future and dependent types are a mainstay and there are specialized math resources for it, but in the present time the PL material will have to do. I simply have nothing better than this.

After dealing with volume 1, I have a basic measure of familiarity with this. As I go along, I will gradually turn that into expertise.

It is going to be a long journey. I just have to buckle up and get through it.

1:40pm. Now let me dug up those cubical type theory links."