"1:10pm.
"The book develops two main conceptual threads:
(1) formal techniques for reasoning about the properties of specific programs (e.g., the fact that a sorting function or a compiler obeys some formal specification); and
(2) the use of type systems for establishing well-behavedness guarantees for all programs in a given programming language (e.g., the fact that well-typed Java programs cannot be subverted at runtime)."

I really am interested in the first one.

Damn, I am going to have to go through this as well.

Well, maybe it won't be as hard as the first part.

1:15pm. "The Coq proofs we're doing now are sufficiently complicated that it is more or less impossible to complete them by random experimentation or "following your nose." You need to start with an idea about why the property is true and how the proof is going to go. The best way to do this is to write out at least a sketch of an informal proof on paper — one that intuitively convinces you of the truth of the theorem — before starting to work on the formal one. Alternately, grab a friend and try to convince them that the theorem is true; then try to formalize your explanation."

1:25pm. The second volume is enormous.

...I am going to be working on this for the next few months at this rate...

Agh, but I do need this. I need these exercises if I am going to do the efficient Bayesian inference algorithm for NNs.

1:30pm. I am going to have to do it. I am going to do vol 2 in Agda and then in Coq, before moving on to volume 3. I guess that will be my goal for the next 3 months.

If I want to do real math with dependent types, then this material is valuable enough as exercise that I have to go through this.

It would be different if it was the future and dependent types are a mainstay and there are specialized math resources for it, but in the present time the PL material will have to do. I simply have nothing better than this.

After dealing with volume 1, I have a basic measure of familiarity with this. As I go along, I will gradually turn that into expertise.

It is going to be a long journey. I just have to buckle up and get through it.

1:40pm. Now let me dug up those cubical type theory links.

https://www.cs.bham.ac.uk/~mhe/HoTT-UF-in-Agda-Lecture-Notes/index.html
Introduction to Univalent Foundations of Mathematics with Agda

https://homotopytypetheory.org/2018/12/06/cubical-agda/
Cubical Agda

1:45pm. https://existentialtype.wordpress.com/2018/01/15/popl-2018-tutorial/

```
https://github.com/RedPRL/redtt

What's the relationship to RedPRL?
The goal of the RedPRL Development Team is to study implementation techniques for higher dimensional type theory, and determine which ideas lead to the most convenient and ergonomic system for developing both higher dimensional mathematics and higher dimensional programming. To this end, our first experiment was the RedPRL Proof Assistant; our second experiment is the redtt proof assistant, which synthesizes what we learned from the experience of building RedPRL and using it to formalize mathematics.
```

```
Unlike structural type theories, a computational type theory is defined not by rules, but by semantics, specifically meaning explanations based on a prior notion of computation.
```

2pm. https://existentialtype.wordpress.com/2011/04/09/persistence-of-memory/#comment-3552

I am reading Harper's blog at the moment.

2:15pm. https://arxiv.org/abs/1611.02108
Cubical Type Theory - a constructive interpretation of the univalence axiom

What I am looking at the moment is a short into to CTT. I want to see a few proofs side by side with the structural one.

2:30pm. The blog post is too thick and does not really clue me in into what is so great about cubical type theory.

Let me give the paper a try. Maybe it will have some motivating examples rather than being replete with jargon. I should also give the HoTT book a try.

2:45pm. This is boring. I am getting distred by /a/ as well.

https://www.youtube.com/watch?v=W5-ulP_JzNc
Cubical Adventures

Let me try watching this for a bit.

2:50pm. https://www.youtube.com/watch?v=zJdwPa_tkSU

It is a waste of time. Let me try this.

2:55pm. https://www.youtube.com/watch?v=hlCVHVtAlqQ

Let me try this.

3:05pm. http://www.redprl.org/en/latest/tutorial.html

The videos and papers are a bust. Let me go through this tutorial.

But first let me take a little break here.

3:40pm. Done with the RedPRL tutorial. Right now there is really not much there.

This stuff is still in the research phase.

https://www.cs.bham.ac.uk/~mhe/HoTT-UF-in-Agda-Lecture-Notes/index.html
Introduction to Univalent Foundations of Mathematics with Agda

Let me go through this. It actually got updated today.

3:45pm. This seems really promising. Let me go for it. I'll dedicate a few days if necessary to going through this."