"4:05pm. https://ncatlab.org/nlab/show/homotopy+type+theory#References
I really should watch some videos by Voevodsky. This feels so weird. Usually dead mathematicians I am hearing about are 100s of years past the expiry date. It is quite a rare honor to receive so many accolades in the present. Even rarer to find a Github repo by them.

4:20pm. https://en.wikipedia.org/wiki/Univalent_foundations

The Wikipedia page has more info on this. My intuition is that I really should make an effort to master this.

4:30pm. 7.6k lines. Yeah, I won't be getting through this in a hurry.

Let me try typechecking the file.

```
   * An empty type [`𝟘`](HoTT-UF-Agda.html#emptytype).

   * A one-element type [`𝟙`](HoTT-UF-Agda.html#onepointtype).
```

These specials symbols are not appearing correctly in Emacs.

```
--------------------------|
abcdefghijklmnopqrstuvwxyz|
ABCDEFGHIJKLMNOPQRSTUVWXYZ|
ᵃᵇᶜᵈᵉᶠᵍʰⁱʲᵏˡᵐⁿᵒᵖ ʳˢᵗᵘᵛʷˣʸᶻ|
ᴬᴮ ᴰᴱ ᴳᴴᴵᴶᴷᴸᴹᴺᴼᴾ ᴿ ᵀᵁⱽᵂ   |
ₐ   ₑ  ₕᵢⱼₖₗₘₙₒₚ ᵣₛₜᵤ  ₓ  |
--------------------------|
----------|
0123456789|
⁰¹²³⁴⁵⁶⁷⁸⁹|
₀₁₂₃₄₅₆₇₈₉|
----------|
------------------------|
αβγδεζηθικλμνξοπρστυφχψω|
ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ|
------------------------|
```

Wow, it seems only the browser gets all the characters right **Edit: No, it does not.**. Just what is it using?

https://wiki.portal.chalmers.se/agda/pmwiki.php?n=Docs.HowToSeeUnicode

Let me try this.

https://github.com/tonsky/FiraCode

Let me give this a shot.

4:50pm. https://stackoverflow.com/questions/40672995/how-to-fix-the-bool-symbol-from-agda-showing-up-as-a-square-box-with-hex-and-dec?noredirect=1&lq=1

Here is a question by me from 3 years ago.

4:55pm. https://plfa.github.io/Fonts/

No, even with the fallback, the characters at the end are not monospaced. Damn.

5pm. Ah, forget it. Let me just go with the Deja Vu Mono + fallback.

At least now with the fallback I have those two math symbols.

```
   * [Universes](HoTT-UF-Agda.html#universes) (types of types), ranged
     over by `𝓤,𝓥,𝓦`.
```

But I do not have these. Agh...

5:15pm. For some reason Times New Roman does not work as a fallback font for me.

Instead of studying the book, I am messing around with fonts. Lame.

5:20pm.

```
(set-fontset-font "fontset-default" nil
        (font-spec :name "Symbola"))
```

Let me try setting Symbola as the default font.

6:15pm. Let me stop here on line 580. I am done for the day.

Yawn. I really am tired at this point. It is always like this whenever I reach a milestone, it takes me a while to shift gears.

6:20pm. Work can only really get done once one has a clear goal in mind.

These 250 pages of notes on univalent foundations in Agda are just the right warmup I need before I start the next SF volume.

I'll check out the references after I am done.

6:25pm. It is pretty amazing that I am studying this at all.

But my experience in programming so far has been unanimous - if one wants to get better at programmer the single best step one could take is to pick a language. When it comes to math that will hold as well.

I will push all the obstacles out of the way and that will make way for the blossoming of my talent in this area.

No matter how much effort it takes. This is the kind of life I've decided upon - for omnipotence everything."