"10:20am.
```
is-set : 𝓤 ̇ → 𝓤 ̇
is-set X = (x y : X) → is-subsingleton (x ≡ y)
```

So this...

10:25am.

```
𝟙-is-set : is-set 𝟙
𝟙-is-set a .a (refl .a) (refl .a) = refl (refl a)
```

Here I had to prove that two propositional equalities are equal.

10:30am.

```
Magma : (𝓤 : Universe) → 𝓤 ⁺ ̇
Magma 𝓤 = Σ \(X : 𝓤 ̇ ) → is-set X × (X → X → X)
```

So with that I can make an interpretation of this.

10:35am. https://stackoverflow.com/questions/56506794/how-to-do-theorems-that-require-mutual-recursion/56534988#56534988

Got a large answer here. I do not feel like accepting it or even studying it at the moment, but it definitely deserves an upvote.

I hadn't expected this problem to be so similar to another not part of SF.

I do not want to admit that I just pared down a SF problem. Oh well, it should be fine.

10:40am. Let me get my focus back on.

Ah, yes, I was intending to study the links next to the Magma.

Let me start with the wiki page and then I will check out the book linked within.

10:50am.

https://books.google.co.uk/books?id=STS9aZ6F204C&pg=PA1&redir_esc=y#v=onepage&q&f=false

This book is under a paywall and does not show for me.

```
Magma 𝓤 = Σ \(X : 𝓤 ̇ ) → is-set X × (X → X → X)
```

I am a bit confused at this definition...

Ah, I got it wrong. is-set only applies to X itself. The magma is in fact a binary function.

10:55am. Now that I think about it `is-set X` should always hold for any type. I mean `(a=b)=(a'=b')` should always be constructible for anything, right?

I am a bit confused as to what that thing is doing then, also why the existential?

11:05am.

```
A function of the underlying sets of two magmas is a called a
homomorphism when it commutes with the magma operations:

\begin{code}
is-magma-hom : (M N : Magma 𝓤) → (⟨ M ⟩ → ⟨ N ⟩) → 𝓤 ̇
is-magma-hom M N f = (x y : ⟨ M ⟩) → f (x ·⟨ M ⟩ y) ≡ f x ·⟨ N ⟩ f y
```

Isn't this distributivity rather than commutativity?

11:20am. Ok...

```
id-is-magma-hom : (M : Magma 𝓤) → is-magma-hom M M (𝑖𝑑 ⟨ M ⟩)
id-is-magma-hom M = λ (x y : ⟨ M ⟩) → refl (x ·⟨ M ⟩ y)
```

I understand the above.

```
is-magma-iso : (M N : Magma 𝓤) → (⟨ M ⟩ → ⟨ N ⟩) → 𝓤 ̇
is-magma-iso M N f = is-magma-hom M N f
                   × Σ \(g : ⟨ N ⟩ → ⟨ M ⟩) → is-magma-hom N M g
                                            × (g ∘ f ∼ 𝑖𝑑 ⟨ M ⟩)
                                            × (f ∘ g ∼ 𝑖𝑑 ⟨ N ⟩)
```

Now what about this?

What does `~` again?

```
_∼_ : {X : 𝓤 ̇ } {A : X → 𝓥 ̇ } → Π A → Π A → 𝓤 ⊔ 𝓥 ̇
f ∼ g = ∀ x → f x ≡ g x
```

11:30am.

```
id-is-magma-iso : (M : Magma 𝓤) → is-magma-iso M M (𝑖𝑑 ⟨ M ⟩)
id-is-magma-iso M = id-is-magma-hom M ,
                    𝑖𝑑 ⟨ M ⟩ ,
                    id-is-magma-hom M ,
                    refl ,
                    refl
```

Let me go through this step by step and then I will take a break.

11:35am. Hmmm, ok I see. By stepping through it I do get a slight feel for it.

The way this feels is kind of like the equivalence stuff from PLFA.

```
is-magma-iso : (M N : Magma 𝓤) → (⟨ M ⟩ → ⟨ N ⟩) → 𝓤 ̇
is-magma-iso M N f = is-magma-hom M N f
                   × Σ \(g : ⟨ N ⟩ → ⟨ M ⟩) → is-magma-hom N M g
                                            × (g ∘ f ∼ 𝑖𝑑 ⟨ M ⟩)
                                            × (f ∘ g ∼ 𝑖𝑑 ⟨ N ⟩)
```

The last two are like `from_to` and `to_from`. That is what it feels like the book is angling for.

11:40am.

```
⌜⌝-is-iso : {M N : Magma 𝓤} (p : M ≡ N) → is-magma-iso M N (⌜ p ⌝)
⌜⌝-is-iso (refl M) = id-is-magma-iso M
```

This part confuses me a little. I am going to have to step through it to make sure that it makes sense.

But let me stop here. I had enough of this for one morning. It is going to take me a bit of effort to internalize this properly."