"3:20pm.
```
_≅ₘ_ : Magma 𝓤 → Magma 𝓤 → 𝓤 ̇
M ≅ₘ N = Σ \(f : ⟨ M ⟩ → ⟨ N ⟩) → is-magma-iso M N f
```

In a way all this passing of existentials is a lot closer to how I did programming in Spiral, that the recent theorem proving stuff.

So I should not be be too taken aback at what is happening.

3:25pm.

```
The following function [will be](HoTT-UF-Agda.html#magmaequivalences) a bijection in the presence of
univalence, so that the identifications of magmas are in one-to-one
correspondence with the magma isomorphisms:

\begin{code}
magma-Id-to-iso : {M N : Magma 𝓤} → M ≡ N → M ≅ₘ N
magma-Id-to-iso p = (⌜ p ⌝ , ⌜⌝-is-iso p )
\end{code}

If we omit the set-hood condition in the definition of the type of
magmas, we get the type of what we could call `∞`-magmas (then the
type of magmas could be called `0-Magma`).

\begin{code}
∞-Magma : (𝓤 : Universe) → 𝓤 ⁺ ̇
∞-Magma 𝓤 = Σ \(X : 𝓤 ̇ ) → X → X → X
\end{code}
```

I really have no idea what to think about about this.

`(then the type of magmas could be called `0-Magma`).` Does he mean the regular magma with the set-hood field?

3:40pm.

```
We could drop the `is-set X` condition, but then we wouldn't get
`∞`-monoids in any reasonable sense. We would instead get "wild
`∞`-monoids" or "incoherent `∞`-monoids". The reason is that in
monoids (with sets as carriers) the neutrality and associativity
equations can hold in at most one way, by definition of set. But if we
drop the set-hood requirement, then the equations can hold in multiple
ways. And then one is forced to consider equations between the
identifications (all the way up in the ∞-case), which is
what "[coherence](https://ncatlab.org/nlab/show/coherence+law) data"
means. The wildness or incoherence amounts to the absence of such
data.
```

```
*Exercise.* Define the type of [groups](https://en.wikipedia.org/wiki/Group_(mathematics)) (with sets as carriers).

*Exercise.* Write down the various types of
[categories](https://en.wikipedia.org/wiki/Category_(mathematics))
defined in the HoTT book in Agda.

*Exercise.* Try to define a type of [topological
 spaces](https://en.wikipedia.org/wiki/Topological_space).
```

Oh boy these exercises.

3:45pm. https://en.wikipedia.org/wiki/Monster_group

Yeah, this stuff is too much for me.

4pm. Yeah, no way I can do these exercises. Using Wikipedia pages as a math reference? Not a great idea.

4:10pm.

```
⁻¹-left∙ : {X : 𝓤 ̇ } {x y : X} (p : x ≡ y)
         → p ⁻¹ ∙ p ≡ refl y
⁻¹-left∙ (refl x) = refl (refl x)
```

I'll admit, at this point all this abstract gibberish is starting to put me to sleep, but I will persevere to understand at least what is written down even if I can't do the exercises.

Let me step through the above function.

4:20pm.

```
\begin{code}
Nat : {X : 𝓤 ̇ } → (X → 𝓥 ̇ ) → (X → 𝓦 ̇ ) → 𝓤 ⊔ 𝓥 ⊔ 𝓦 ̇
Nat A B = (x : domain A) → A x → B x
\end{code}

We don't need to specify the naturality condition, because it is
automatic:

\begin{code}
Nats-are-natural : {X : 𝓤 ̇ } (A : X → 𝓥 ̇ ) (B : X → 𝓦 ̇ ) (τ : Nat A B)
                 → {x y : X} (p : x ≡ y)
                 → τ y ∘ transport A p ≡ transport B p ∘ τ x
Nats-are-natural A B τ (refl x) = refl (τ x)
\end{code}
```

Boredom is really hitting me hard as I read this.

4:25pm.

```
transport A p a ≡ b
```

I wonder if this related to that type inference issue in Agda when it comes to implicit arguments?

4:30pm. I had enough. Let me just skim the rest of the book and then I'll take a look at some of the references at the beginning.

```
Provide definitions for the following type declarations:

\begin{code}
joyal-equivs-are-invertible : {X : 𝓤 ̇ } {Y : 𝓥 ̇ } (f : X → Y)
                            → is-joyal-equiv f → invertible f

joyal-equivs-are-equivs : {X : 𝓤 ̇ } {Y : 𝓥 ̇ } (f : X → Y)
                        → is-joyal-equiv f → is-equiv f

invertibles-are-joyal-equivs : {X : 𝓤 ̇ } {Y : 𝓥 ̇ } (f : X → Y)
                             → invertible f → is-joyal-equiv f
```

4:35pm. I am done with this book. There is nothing of interest in here to me.

https://video.ias.edu/voevodsky14
Univalent Foundations: New Foundations of Mathematics

Let me watch this.

https://youtu.be/zw6NcwME7yI?t=1065

"And it was clearly not an accident. A technical argument by a trusted advisor, which is hard to check and looks similar to arguments known to be correct, is hardly ever checked in detail."

The whole talk so far has been on mathematical errors.

5:05pm. https://youtu.be/zw6NcwME7yI?t=1238

"I can see two factors that contributed to this outrageous situation: (of an error persisting for decades)

- Simpson claims to have constructed a counterexample, but he was not able to show where in out paper the mistake was. Because of this, it was not clear whether we made a mistake somewhere in our paper or he made a mistake somewhere in his counterexample.

- Mathematical research currently relies on a complex system of mutual trust based on reputations. By the Simpson's paper both Kapranov and I had strong reputations. Simpson's paper created doubts in our results which led to it being unused by other researchers, but **no one came came forward and challenged us on it**. And that includes Simpson himself."

https://youtu.be/zw6NcwME7yI?t=1950

"Today, only a few years later computer verification of proofs and of mathematical reasoning in general looks completely practical to many people who work on Univalent Foundations and Homotopy Type Theory.

The roadblock that prevented generations of interested mathematicians and computer scientists from solving the problem of computer verification of mathematical reasoning was the unpreparedness of foundations of mathematics for the requirements of this task."

6:15pm. Let me stop on line 2300.

Though boredom hit me like a brick I am starting to realize something - univalent foundations are not like a whole new language, but rather more like a new set of techniques + a new axiom.

Hence rather than aiming for 70 pages of the book per day, I should instead aim for 10 and make it a slow study. I should give myself time to I should just be done with this by the time the half yearly review time rolls around.

I'll take it easy and do these simple boring functions one at a time.

I never really studied math to this level of depth before, so I should try and take it seriously even though it is putting me to sleep. Eventually I will acclimate to this and that will be my first real step into the world mathematics.

6:25pm. Now, let me get to slacking. I had enough for one day. I know that I had an outsized break today and I have barely put in any work at all. I'll strive to do better from here on out.

Somehow the talk by Vladimir was motivating. It was both on avoiding mistakes and the need for computed aided reasoning. We are the like of mind when we think that proof assistants will inevitably be a mainstay of mathematics.

Rest in peace."